SPO-6131 Move widget in file
