+++ examples/cat_facts/lib/ui/screen/facts/facts_screen.dart
@@ -14,6 +14,7 @@
 
 import 'package:cat_facts/data/facts/fact.dart';
 import 'package:cat_facts/data/theme/app_theme.dart';
+import 'package:cat_facts/ui/screen/facts/components/fact_list_tile.dart';
 import 'package:cat_facts/ui/screen/facts/facts_screen_wm.dart';
 import 'package:cat_facts/ui/screen/facts/theme_button.dart';
 import 'package:flutter/material.dart';
@@ -95,51 +96,4 @@ class _FactsScreenState
   }
 }
 
-class FactListTile extends StatelessWidget {
-  const FactListTile({
-    Key? key,
-    required this.wm,
-    required this.el,
-    required this.position,
-  }) : super(key: key);
 
-  final FactsScreenWidgetModel wm;
-  final Fact el;
-  final int position;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
-      child: Column(
-        children: [
-          Row(
-            children: [
-              Text(
-                'Fact $position',
-                style: const TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-              const SizedBox(width: 4),
-              StreamBuilder<AppTheme?>(
-                stream: wm.currentTheme(),
-                builder: (context, snapshot) {
-                  return Expanded(
-                      child: Divider(
-                    color: (snapshot.data == AppTheme.light)
-                        ? Colors.black
-                        : Colors.white,
-                  ));
-                },
-              )
-            ],
-          ),
-          const SizedBox(height: 10),
-          Text(el.fact ?? ''),
-        ],
-      ),
-    );
-  }
-}
